lib.types: fix loaOf behavior of long lists
Assigning a list of 10 or more elements to an option having the type
`loaOf a` produces a configuration value that is not honoring the
order of the original list. This commit fixes this and a related issue
arising when 10 or more lists are merged into this type of option.

(cherry picked from commit d7903cb2434b01873ba344b94c1a901482b744d5)