firefox: allow overriding config of wrapped package
By moving the `cfg` variable into the wrapper arguments we are able to
override it for an already wrapped package. For example, with this
change one can have

    pkgs.firefox-devedition-bin.override {
      cfg.enableBrowserpass = true;
    }

which would otherwise be difficult to accomplish for packages having a
complicated wrapped definition in `all-packages.nix`.

(cherry picked from commit 06d8666a60174dd3083d1516f9f405251601f476)