lib/types: allow custom `submoduleWith` descriptions
Currently the only way to set the description for a submodule type is to
use `freeformType`. This is not ideal as it requires setting a
freeform type, and evaluates the submodule config unnecessarily.

Instead, add a `description` argument to `submoduleWith`.

(cherry picked from commit ceebdcfc2ca2d37872c46f224503a36f4d25daa9)